+++ apps/meteor/tests/e2e/page-objects/admin.ts
@@ -20,8 +20,16 @@ export class Admin {
 		return this.page.locator('[role="link"]', { hasText: name });
 	}
 
+	getUserRow(username?: string): Locator {
+		return this.page.locator('[role="link"]', { hasText: username });
+	}
+
 	get btnSave(): Locator {
 		return this.page.locator('button >> text="Save"');
+	}
+
+	get btnEdit(): Locator {
+		return this.page.locator('button >> text="Edit"');
 	}
 
 	get privateLabel(): Locator {
@@ -197,7 +205,19 @@ export class Admin {
 		return this.page.locator('input[placeholder="Room"]');
 	}
 
+	get inputUsers(): Locator {
+		return this.page.locator('input[placeholder="Users"]');
+	}
+
+	get btnAdd(): Locator {
+		return this.page.getByRole('button', { name: 'Add' });
+	}
+
+	get btnBack(): Locator {
+		return this.page.getByRole('button', { name: 'Back' });
+	}
+
 	getUserRowByUsername(username: string): Locator {
-		return this.page.locator('tr', { hasText: username })
+		return this.page.locator('tr', { hasText: username });
 	}
 }